backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pdfplumber>=0.11.0
pydantic>=2.6.4
python-dotenv>=1.0.1
openai>=1.14.0
supabase>=2.4.5
apscheduler>=3.10.4
python-multipart>=0.0.9
pymupdf>=1.24.0
pillow>=10.2.0
python-doctr[torch]>=0.9.0
